packages/core/pyproject.toml
[project]
name = "os_ai_core"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
]

# runtime deps
dependencies = [
  "injector>=0.21.0",
  "pyautogui>=0.9.54",
  "Pillow>=10",
  "pyperclip>=1.8",
  # internal packages
  "os_ai_llm>=0.1.0",
  "os_ai_llm_anthropic>=0.1.0",
  "os_ai_llm_openai>=0.1.0",
  "os_ai_os>=0.1.0",
  "os_ai_os_macos>=0.1.0",
]

[tool.uv]
package = true
